requirements.txt
llama-cpp-python>=0.2.20
psutil>=5.9.0
pywin32>=306
python-dotenv>=1.0.0
requests>=2.31.0
sentence-transformers>=2.2.2
rich>=13.0.0
google-generativeai>=0.4.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
